+++ framework/src/main/java/org/smartfrog/services/anubis/partition/test/controller/ControllerConfiguration.java
@@ -48,12 +48,12 @@ public class ControllerConfiguration {
         return 1;
     }
 
-    public long heartbeatInterval() {
-        return 2000L;
+    public int heartbeatInterval() {
+        return 2000;
     }
 
-    public long heartbeatTimeout() {
-        return 3L;
+    public int heartbeatTimeout() {
+        return 3;
     }
 
     @Bean
